OpenMP 2.5 comformance fixes

+++ util/mkcromfs.cc
@@ -637,7 +637,7 @@ namespace cromfs_creator
         MutexType blockify_lock;
 
         #pragma omp parallel for schedule(guided)
-        for(size_t p=0; p<collection.size(); ++p) // Step 4.
+        for(long p=0; p < (long)collection.size(); ++p) // Step 4.
         {
             direntry& ent = collection[p];
             const struct stat64& st = ent.st;
